refactor ipc getAvailableWallet
